Extract the bulk of buildSeriesInfoFragment
